power manager: list actions only if is user allowed to call it

+++ libraries/razorqt/powermanager.cpp
@@ -89,21 +89,33 @@ QList<QAction*> PowerManager::availableActions()
     QAction * act;
 
     // TODO/FIXME: icons
-    act = new QAction(XdgIcon::fromTheme("system-suspend-hibernate"), tr("Hibernate"), this);
-    connect(act, SIGNAL(triggered()), this, SLOT(hibernate()));
-    ret.append(act);
-    
-    act = new QAction(XdgIcon::fromTheme("system-suspend"), tr("Suspend"), this);
-    connect(act, SIGNAL(triggered()), this, SLOT(suspend()));
-    ret.append(act);
+    if (m_upower->canHibernate())
+    {
+        act = new QAction(XdgIcon::fromTheme("system-suspend-hibernate"), tr("Hibernate"), this);
+        connect(act, SIGNAL(triggered()), this, SLOT(hibernate()));
+        ret.append(act);
+    }
 
-    act = new QAction(XdgIcon::fromTheme("system-reboot"), tr("Reboot"), this);
-    connect(act, SIGNAL(triggered()), this, SLOT(reboot()));
-    ret.append(act);
+    if (m_upower->canSuspend())
+    {
+        act = new QAction(XdgIcon::fromTheme("system-suspend"), tr("Suspend"), this);
+        connect(act, SIGNAL(triggered()), this, SLOT(suspend()));
+        ret.append(act);
+    }
 
-    act = new QAction(XdgIcon::fromTheme("system-shutdown"), tr("Shutdown"), this);
-    connect(act, SIGNAL(triggered()), this, SLOT(halt()));
-    ret.append(act);
+    if (m_upower->canReboot())
+    {
+        act = new QAction(XdgIcon::fromTheme("system-reboot"), tr("Reboot"), this);
+        connect(act, SIGNAL(triggered()), this, SLOT(reboot()));
+        ret.append(act);
+    }
+
+    if (m_upower->canHalt())
+    {
+        act = new QAction(XdgIcon::fromTheme("system-shutdown"), tr("Shutdown"), this);
+        connect(act, SIGNAL(triggered()), this, SLOT(halt()));
+        ret.append(act);
+    }
 
     act = new QAction(XdgIcon::fromTheme("system-log-out"), tr("Logout"), this);
     connect(act, SIGNAL(triggered()), this, SLOT(logout()));
